Update to draft 19
Even though a device id is now 128 bits, spnfs is only concerned with
the rightmost 64 bits for now.  The leftmost 64 bits appear to be destined
to become an fsid.  The generic pnfs layer is hardcoding this to 0 for the
moment.  Wait and see how this plays out before deciding what needs to be
done in spnfs.

Signed-off-by: Mike Sager <sager@netapp.com>