pnfs: initialize pnfs pageio for write
In 2.6.18.3, there are several write I/O paths through the client code.

In 2.6.24, there is only one write I/O path through the client code.  The
interface to the nfs page cache has been re-written. It is now shared between
the read and write code paths, and moved entirely to fs/nfs/pagelist.c.
The pageio is setup in the nfs address_space operations, nfs_readpages and
nfs_writepages. We need to determine if we are performing pNFS or MDS I/O
the pageio code so that we coalese correctly.

The read path (nfs_readpages) can calculate the readahead length and use it
to determine I/O threshold.

The write path (nfs_writepages) receives an address_mapping of all the
outstanding writes to the server (possibly via multiple opens).

Add some housekeeping to struct nfs_pageio_descriptor and some write path
code to the pg_test routine to set the pg_bsize to wsize (to MDS) or
ds_wsize (to pNFS) during page coalescing.

Signed-off-by: Andy Adamson<andros@umich.edu>
[pnfs: always initialize pnfs pgio desc fields in pnfs_pageio_init_write]
    currently, when pnfs is not used the pnfs specific fields in nfs_pageio_descriptor
    aren't initialized.
Signed-off-by: Benny Halevy <bhalevy@panasas.com>