pnfs: initialize pageio for read
The page cache can already have data in the readahead range. The nfs_readpages
pagelist will 'skip' these pages.  In this case using the number of pages in
the pagelist to calculate the layout range will be incorrect.

Use the first and last page in the pagelist to calculate the layout request
range.

The pageio coalesce code will create requests to fill the holes of data not
present in the page cache.

So! We could delay setting the rsize (and whether to use pNFS or RPC to MDS)
until the coalesce code, sending requests below the threshold to the MDS....

This code uses the whole readahead range for the threshold decision, as is
done in 2.6.18.3.

This patch calculates the readahead range, tries to get a layout, and if
successful, sets the coalesce size to ds_rsize.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>

[pnfs: prevent offset overflow in readahead_range]
    wb_index needs to be typecasted to u64 or loff_t before shifting left
    with PAGE_CACHE_SHIFT to prevent overflow on 32-bit architectures
    since pgoff_t is unsigned long.
[pg_test changes to pnfs_set_ds_size]
[extracted from: pnfs: use boundary in coalese page cache request]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[was: pnfs: read path set ds_size]
    Use the pNFS ds_rsize for read pageio.

    When using the nfs page cache for pNFS I/O, it is necessary to setupe pages
    using the rsize of the storage device instead of the rsize of the MDS.

    In 2.6.18.3, there are several read I/O paths through the client code and
    the rsize was used in each of these paths to setup the nfs page cache. Since
    several functions used the rsize, an rpc_ops->rsize function makes sense.

    In 2.6.24, there is only one read I/O path through the client code. The
    interface to the nfs page cache has been re-written. It is now shared between
    the read and write code paths, and moved entirely to fs/nfs/pagelist.c.

    As a result, there is now only one place to set the rsize.

    This patch removes the rpc_ops->rsize function pointer and
    sets the rsize to the pNFS ds_size when pNFS is being used.

    pNFS is not being used if layoutget fails, or the count is below the
    threshold.
Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: fix compile problems if CONFIG_PNFS turned off - read.c]
Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
[pnfs: client layout_cache: pnfs_update_layout returns an lseg]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>