pnfs: pnfs_return_layout stateid processing
Add a done routine to the async layout return call and update the
stateid.

For RETURN_FILE, we need to pass the stateid we received by
CB_LAYOUTRECALL down to LAYOUTRETURN.
This patch adds an optional stateid pointer on the pnfs return_layout.

FIXME: We either store the stateid somewhere else, or get a reference to the
pnfs_layout_type for all recall types, not just for RECALL_FILE.

Signed-off-by: Andy Adamson<andros@netapp.com>
[pnfs: carry reference to pnfs_layout_type in nfs4_pnfs_layoutreturn]
[pnfs: pass stateid to return_layout]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: client could ignore some CB_LAYOUTRECALLs]
Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>