pnfs: get a layout in the pnfs fileoperations write method
We need to get a layout prior to coalesing pages because if we can't get a
layout, we should coalesce the pages for MDS write. If we can get a layout,
for the file layout module it makes sense to look up and store the dataserver
while coalsecing, because if we can't get a dataserver, we can coalesce the
pages for MDS write.

In the read path, we can try to get a layout in pnfs_pageio_init_read()
called just prior to page coalescing. In the write path, we don't have the
offset nor count at pnfs_pageio_init_write.

Declare a pnfs file operations structure with a pnfs write method,
pnfs_file_write() and try to get a layout for the request range.

Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: Add file layout comments and debugging info.]
    Other than comments, the main change is adding the
    string version of the ip addr to the device so
    it can be printed out for easier debugging.
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: refactor pnfs call data out of nfs_{read,write}_data]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: dynamically initialize pnfs_v4_clientops]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[remove CONFIG_PNFS]
Signed-off-by: Andy Adamson <andros@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>