pnfs: HACK: modify write_end_cleanup
This needs to be changed, but will require a major rewrite
of the block layout's IO code.  Including it here so I can
get some current code into the tree.

Change write_end_cleanup API to allow block driver the ability to
handle server blocks larger than a page in size.  It needs to
initiate writes to adjacent pages, so pass in filp so that it
can call the appropriate aop functions.  A field is also added to
fsdata to hold the list of pages.

Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>