pnfs: pnfs_call_done
pnfs_call_done implements the common cleanup code for the read, write,
and commit call completion functions.

The rpc_release phase happens outside the rpc execute core loop
since it may (and does in our case) release the task structure itself
which is embedded in the call data.
rpc_free_task that calls rpc_release is being called either
from rpc_put_task in the sync rpc case as the last thing to happen on
task or, in the async case, via the workqueue.

All the pnfs _done function needs to do is to call the original
rpc_call_done function (it is complicated to have rpc_execute
do that for us since we're already being called inside rpc_exit_task)
and restore the original tk_ops which the rpc engine code will use
to call the rpc_release method later on.

Since on PNFS_NO_RPC path there's no rpc task running and the
_done function does not return to the rpc layer, we need to call
the rpc_release method ourselves.

[extracted from: pnfs: refactor pnfs read, write, and commit done functions]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: no rpc nfs page cache cleanup]
Signed-off-by: Andy Adamson<andros@umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: don't call rpc_release from pnfs_call_done on rpc completion]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: rename -pnfs_rpc_done to pnfs_call_done]
    Since it is called back also from non-rpc paths.
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: put_lseg in pnfs_call_done]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: compress struct pnfs_call_data]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>