nfs41: state reclaimer renew lease error handling
The state reclaimer has been called due to a NFS4_STALE_CLIENTID or NFS4_EXPIRE
error. First it tries to renew the old lease (network partion case) which for
v4.1 means an nfs4_proc_sequence call. Any error from the sequence operation
means that the lease cannot be renewed.

Switch on the minorversion to handle the error return from the renew lease call.

Signed-off-by: Andy Adamson<andros@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>