pnfs: getdevicelist and getdeviceinfo
proc and xdr code for getdevice{list,info}

The getdeviceinfo call is made by the client when a layoutget operation
returns an unknown device id.
Use the filehandle of the layoutget call as the current filehandle
for the getdeviceinfo call, instead of the filehandle of the pNFS mount point.

The 2.6.18.3 getdeviceinfo call used the root filehandle of the mount point
stored in struct nfs_server->fh. The 2.6.24 nfs client no longer saves the
root filehandle. Instead, a local variable is used.

There is no reason to use the MDS mountpoint filehandle for getdeviceinfo.
Instead of reverting the 2.6.24 code to save the filehandle of the pNFS
mount point, this patch uses the filehandle of the layoutget call that
initiated the getdeviceinfo call.

Note: If no notifications are requested represent gdia_notify_types as an empty
bitmap array rather than one consisting of a single zeroed entry.
Thanks to Dean Hildebrand for suggesting this optimization
and to Peter Staubach for convincing that it's worth it.

page based decoding for getdeviceinfo:
    First try with one page for the result. If -NFS4ERR_TOOSMALL and returned
    mincount fits with in NFS4_GETDEVINFO_MAXSIZE, retry.

    If using more than one page, use vmap to map a contiguous memory portion for
    decoding. The use of vmap to set contiguous memory for decoding OK because
    getdeviceinfo is rarely called.

[extracted from: pnfs: Initial check-in of pNFS File Layout Driver.]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: getdeviceinfo]
Signed-off-by: Andy Adamson<andros@umich.edu>
[pnfs: fix nfs4_call_sync for nfs4_pnfs_getdeviceinfo]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: Define 128 bit deviceid type for client]
Define 128 bit deviceid for common pNFS and 64bit maj/min
for file layout
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
[pnfs: Cleanup layout driver device callbacks.]
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
[pnfs: Move getdeviceinfo to draft-19 on client]
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
[pnfs: encode empty getdeviceinfo notify bitmap rather than zeroed]
[pnfs: do not send putfh for getdeviceinfo]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs: Move getdevicelist client to draft 19]
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
[pnfs: use new client xdr: getdevice{list,info}]
Signed-off-by: Andy Adamson<andros@netapp.com>
[pnfs: simplify getdeviceinfo xdr structures]
[pnfs: remove unused file handle arg from getdeviceinfo]
[pnfs: rename getdevicelist maximum size]
[pnfs: change maximum getdeviceinfo size]
Signed-off-by: Andy Adamson <andros@netapp.com>
[pnfs: page based decoding for getdeviceinfo]
Signed-off-by: Andy Adamson <andros@netapp.com>
[pnfs: Return info about length of opaque in GETDEVICEINFO]
Signed-off-by: Fred Isaman <iisaman@citi.umich.edu>
[pnfs: fix decode_getdevicelist decoding of cookie]
[pnfs: fix encode_getdevicelist's xdr Xcoding pointer type]
[pnfs: clean up xdr]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfs-submit: use the session max response size for getdeviceinfo's maxcount]
    Per Trond, no need to try a small maxcount first.  Base the maxcount
    on the session's max response size.
    Tested multi-page getdeviceinfo responses using the pyNFS server by
    creating big stripe index lists.
[pnfs: in get_device_info, free page pointers]
Signed-off-by: Mike Sager <sager@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>