nfsd41: separate clientid slot from session slot
The nfs41 single slot clientid cache holds an encoded create session response
which has a maximum size of 88 bytes. The slot does not need the inuse,
cachethis or other fields that the multiple slot session cache uses.

Declare a new structure with a static buffer to struct nfs4_client to cache
the encoded create session response.

Signed-off-by: Andy Adamson <andros@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>