pnfs: free layout segments only after returning them
The layout segment instances are needed for the obj layout driver
to encode LD specific layout return information.

This is also needed as part of the layout get/return synchronization feature
where any operation that requites a layout in a range conflicting with
one that's being returned will wait until the layout return completes.
Then, the old segments will be unhashed and freed and new layout gets
to these ranges can be resumed.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>