pnfsd: ds state
When a DS verifies a state id on the MDS, the
list of delegation stateids also needs to be
searched.

export ops:
cb_get_state: callback from fs on MDS only
get_state: call fs on DS only
cb_change_state: callback from fs on DS only

[extracted from pnfsd: Initial pNFS server implementation.]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: Check for deleg state on MDS from DS.]
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: Use 128 bit deviceid on server]
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: remove unused symbol exports]
Signed-off-by: Marc Eshel <eshel@almaden.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: convert struct pnfs_get_state.ino from unsigned long to u64]
Signed-off-by: David M. Richter <richterd@citi.umich.edu>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: Add new mutex to protect ds lock state.]
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: release ds stateid]
Signed-off-by: Marc Eshel <eshel@almaden.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: include linux/sched.h from nfs4pnfsds.c]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: Fix stateid handling on DS.]

This is the first major patch that fixes stateid and lock
handling on the DS.
The major problem is deadlocks due to DS<->MDS communication while
holding the state lock.

This patches fixes the problem by unlocking the state lock
where necessary and protecting the ds state structures using
a new ds_state_lock.

The problem is that nfsv4_ds_get_state holds
the ds state lock while verifying the stateid on the mds.
To allow the DS to release the lock while calling the mds
and avoid a barrage of stateid verifications to the mds
for the same stateid, only the first request calls
the mds while the others wait.  Once the first request
has verified the stateid, it wakes up the other nfsd threads
waiting on the bit for this stateid.  To ensure only a single
request is sent to the mds, before the first request calls
the mds, it adds a stateid to the list while setting it
to be uninitialized.  Other threads that see an uninitialized
stateid know they must wait until it is initialized or put
in an error state.  If the stateid cannot be verified, then
it is put in an error state and removed from the list (using a kref).

[pnfsd: Use ds_lock_state for change state on ds.]
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: beef up debug prints for ds stateid functions]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: Fix memory leak by addings refs to client and mds ds structs.]
Signed-off-by: Dean Hildebrand <dhildeb@us.ibm.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[pnfsd: use STATEID_FMT and STATEID_VAL for printing stateids]
[pnfsd: filelayout: fix possible NULL deref in nfsv4_ds_get_state]
[pnfsd: move vfs api structures to nfsd4_pnfs.h]
[pnfsd: filelayout: convert to using new pnfs api]
[pnfsd: filelayout: file layout specific callbacks]
[pnfsd: get rid of generic use of old cb ops in export_operations]
[pnfsd: obliterate old vfs api]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>
[Move pnfsd code from nfs4state.c to nfs4pnfsd.c]
Signed-off-by: Andy Adamson <andros@netapp.com>
[pnfsd: filelayout: Use fh_copy_shallow make code simply]
Signed-off-by: Bian Naimeng <biannm@cn.fujitsu.com>
[pnfsd: expire_client code cleanup]
Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
[pnfsd: fix compiler warnings when CONFIG_PNFSD is not defined]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>