pnfs-obj: objio_osd device information retrieval and caching
When a new layout is received in objio_alloc_lseg all device_ids referenced
are retrieved. The device information is queried for from MDS and then the
osd_device is looked-up from the osd-initiator library. The devices are
cached in a per-mount-point list, for later use. At unmount all devices are
"put" back to the library.

TODO: Only single device per-layout is implemented, multi-device layouts will
      come later at a separate patch.

TODO: The device cache can get big. Cap it's size. Keep an LRU and start to return
      devices which were not used, when list gets to big, or new entries allocation
      fail.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>