pnfs-obj: objlayout_encode_layoutreturn Implementation.
An io_state pre-allocates an error information
structure for each possible osd-device that might error
during IO. When IO is done if all was well the io_state
is freed. (as today). If the I/O has ended with an
error, the io_state is queued on a per-layout err_list.
When eventually encode_layoutreturn() is called, each error
is properly encoded on the XDR buffer and only then the
io_state is removed from err_list and de-allocated.

It is up to the io_engine to fill in the segment that fault
and the type of osd_error that occurred. By calling
objlayout_io_set_result() for each device.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>