pnfsd: set_device_notify
Set the device notify_types at GETDEVICEINFO time.
The call may be used to initially set or to update
the requested notify types for the given deviceid.
The implementor of the method must update the notify_types
member with the supported notification types.
If none are supported, this method can be left unimplemented.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>