FIXME: spnfs: don't check for existence of pnfs export op functions
Maybe SQUASH with 303a886f

Apparently, the export code now makes a second pass through check_export().
This results in exportfs failing in an spnfs configured environment.  For
example, if CONFIG_SPNFS is enabled and we're a DS, s_pnfs_op will get
set to spnfs_ds_export_ops.  Second time through, this will fail the check
for existence of pnfs export op functions since s_pnfs_op will be non-NULL
and no MDS pnfs functions are defined (since we're a DS).

Since this check seems to be in place for underlying filesystems that
define pnfs operations and not spnfs, we can just skip it.

FIXME: bhalevy: That's fine for now, but ideally I'd prefer to have a
similar check for a DS-provided vector.

Signed-off-by: Mike Sager <sager@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>