SQUASHME: pnfs-submit: call layoutcommit after flushing inode's data to disk.
This is for a bug introduced to 2.6.34. In 2.6.32 and 2.6.33 we call
layoutcommit in nfs_sync_mapping_wait(), but in 2.6.34 we use
sync_inode() to sync inode's data, so the layoutcommit code is gone.

BTW: In current code, layoutcommit_ctx will increase refcount of
nfs_inode's ctx, so if layoutcommit_ctx is not NULL, we could not
reach nfs4_close_context ... --> __nfs_close().
So pnfs_layoutcommit_inode() in __nfs_close() will not be called in
whatever situation.
Why we have to use nfs_inode's ctx as layoutcommit_ctx, since we
only need its rpc_creds actually?

Boaz:
  Without this patch non-files layouts which return NFS_FILE_SYNC
  from writes, will eventually hang on uncommitted inodes.

  This patch reveals a bad design pattern from the pnfs-client. On
  none-files layouts nfs_post_op_update_inode_force_wcc or friends
  are never called and i_size_write() is never preformed on client.
  What happens is that the layoutcommit eventually updates the server.
  In Linux the sever would update i_size_attr and mtime. It would then
  be detected by the client and i_size is fetched from server. This is
  not nice client behaviour, and is not done so for files and none-pnfs
  IO. Sigh!
  I think also files-lo might suffer from this in some cases, which means
  if layoutcommit is not preformed (It's just an hint) then writeout will have
  i_size  corruption.

Signed-off-by: Tao Guo <guotao@nrchpc.ac.cn>
Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
[s/do_layoutcommit/layoutcommit_needed/]
Signed-off-by: Benny Halevy <bhalevy@panasas.com>