DEVONLY: pnfs: Add -pnfs suffix to EXTRAVERSION
The official way of adding a tree identifier tag to the Kernel
version-string is a presence of a localversion-* file next to
the Makefile. It's content is appended to the Linus supplied
version-string from Makefile.

After this patch the output is the same as before xxx-2.6.35-rc2-pnfs
But is a more comfortable way to manage it.

(The appended "+" in post 2.6.35-rc Kernels is not resolved with
 this. But I have sent a patch to Linus to fix this issue)

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>