FIXME: pnfsd: do not expire client on cb_layout error.
This may lead to deadlocks on the nfsd state lock.
Instead, wait for the laundromat to expire the client.

TODO: Regardless, even if the callback succeeded, we need to track
a list of recalled layout and maintain it as they returned by the client
Then, scrub it in the laundromat and expire the client if it hasn't
returned all recalled layouts in a timely manner (e.g. after 2
lease periods after its last return with the respective stateid)

Signed-off-by: Benny Halevy <bhalevy@panasas.com>