SQUASHME pnfs-submit fix rcu bug in pnfs deviceid cache
Fix a bug where an invalid (freed) deviceid node could be returned by
nfs4_find_deviceid.

Instead of taking a reference on the deviceid node after the find or add
in nfs4_set_layout_deviceid, take a reference to the deviceid node before
giving up the rcu_read lock in nfs4_find_devicei or the spin_lock in
nfs4_add_deviceid.

Use atomic_inc_not_zero so as not to hold a reference on an element which
has been deleted from the hlist.

Add comments.

Reported-by: Fred Isaman <iisaman@netapp.com>
Signed-off-by: Andy Adamson <andros@netapp.com>
Signed-off-by: Benny Halevy <bhalevy@panasas.com>