NFS/VFS: Fix the nfs sillyrename regression
Commit 949854d0 (fs: Use rename lock and RCU for multi-step operations)
broke sillyrename on NFS because it NULLs out the dentry->d_parent
prior to calling dentry->d_op->iput()

The following patch adds a new callback that is used by d_kill() to
notify the filesystem that the sillyrenamed dentry is no longer in use,
and that it may be unlinked. It takes the parent dentry and negative
dentry as its argument, and is used by the NFS layer to complete
the sillyrename.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>