FIXME: spnfs-block: restore i_op->fallocate
spnfsd-blocks needs the old inode_operations API for fallocate
as it does not have a struct_file in hand.

As all file systems (but xfs) currently use the struct file argument
to get to the inode move their implementation back into a inode operation.
Introduce generic_file_fallocate that can be used as the file_operations
method that just does that and calls i_op->fallocate.

Refactor the xfs implementation and introduce _xfs_vn_fallocate
that takes an addition attr_flags, which value depends on the struct file
argument to xfs_file_fallocate.

On 2011-03-30 17:58, Christoph Hellwig wrote:
> NAK.  Not only isn't spnfsd-block not upstream, but I probably never will be
> given what a piece of junk it is.
>
> Second making fallocate a file operation was done on purpose, and all the
> other filesystem need the same fix that xfs has - making the allocation
> stable if done on an O_SYNC file descriptor.

Signed-off-by: Benny Halevy <bhalevy@panasas.com>
NACK-by: Christoph Hellwig <hch@lst.de>