pnfsd: Return all ROC layouts On Close
As talk to with Trond, when a client has sent an nfs4_close, the
server can surly assume that the given client has already forgotten
all it's ROC given layouts. And in fact the current Linux-client goes
to grate length to make sure all ROC layouts are invalidated, before
sending the close. If a new open+IO comes in and tries to get a new layout,
it will wait for the nfs4_close to return before sending the new layout_get.

So in light of above the server should simulate layout_return and remove
any layouts marked ROC on nfs4_close.

Current code is protocol correct, but a more fine-grained approach can be
devised to also release layouts in open_downgrade, of all unused iomodes.
But to do this I'll need the help of an expert.

With this patch in, I'm finally able to run with the new Linux-Client together
with a Linux-Server. Before, the layouts where only released and freed on umount
and in a modest machine like the UML I use, few 10s of "git checkout linux" the
server would oom.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
[remove iomode parameter]
[moved call to pnfs_roc before release_open_stateid]
Signed-off-by: Benny Halevy <benny@tonian.com>