SQUASHME: pnfsd: Don't call fs_layout_return on a none-file
If we had a layout on the file the nfs4_file would be referenced and we should
have found it. Since we don't then it means all layouts were ROC and at this
point we returned all of them on file close.

So why does the Linux wonderful forgetfull client returnes it's ROC layouts on
inode evict() (I hit this on umount at client). Actually this is on me, because
it is a bug in Generic client code, but can only manifest with pnf-obj after an
IO error.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Signed-off-by: Benny Halevy <bhalevy@tonian.com>