pnfsd-exofs: Two clients must not write to the same RAID stripe
If we have file redundancy  RAID1/4/5/6 then two clients cannot
write to the same stripe/region.

We take care of this by giving out smaller regions of the file.
Before any layout_get we make sure to recall the same exact
region from any client. If a recall was issued we return
NFS4ERR_RECALLCONFLICT. The client will come again later for
it's layout.

Meanwhile the fist client can flush data and release the
layout. The next time the segment might be free and the
lo_get succeed.

It is very possible that multiple writers will fight and
some clients will starve forever. But the smaller the
region, and if the clients randomize a wait, it should
statistically be OK.
(We could manage a fairness queue. What about an lo_available
 notification)

On the other hand a very small segment will hurt performance,
so default size is now set to 8 stripes.
TODO: Let segment size be set in sysfs.

TODO:
  For debugging we always give out small segments. But we should
  only start giving out small segments on a shared file. The
  first/single writer should get a large seg as before.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Signed-off-by: Benny Halevy <bhalevy@tonian.com>