ARM: iop13xx: remove duplicate IOP13XX_PCIE_LOWER_IO_BA
iop13xx ended up with two conflicting definitions for
IOP13XX_PCIE_LOWER_IO_BA, where one of them is used to set up
the hardware window and should be zero, while the other one is
used to set the location of the second mapping into the virtual
address space.

This kills the second definition and hardcodes the 64KB offset
that is already hardcoded in the bios32.c file now.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>